refactor: move existing advisory defs to v1 package
