+++ pkg/advisory/create.go
@@ -4,13 +4,13 @@ import (
 	"fmt"
 
 	"github.com/wolfi-dev/wolfictl/pkg/configs"
-	"github.com/wolfi-dev/wolfictl/pkg/configs/advisory"
+	v1 "github.com/wolfi-dev/wolfictl/pkg/configs/advisory/v1"
 )
 
 // CreateOptions configures the Create operation.
 type CreateOptions struct {
 	// AdvisoryCfgs is the Index of advisory configurations on which to operate.
-	AdvisoryCfgs *configs.Index[advisory.Document]
+	AdvisoryCfgs *configs.Index[v1.Document]
 }
 
 // Create creates a new advisory in the `advisories` section of the configuration
@@ -29,10 +29,10 @@ func Create(req Request, opts CreateOptions) error {
 
 	case 1:
 		// i.e. exactly one advisories file for this package
-		u := advisory.NewAdvisoriesSectionUpdater(func(cfg advisory.Document) (advisory.Advisories, error) {
+		u := v1.NewAdvisoriesSectionUpdater(func(cfg v1.Document) (v1.Advisories, error) {
 			advisories := cfg.Advisories
 			if _, existsAlready := advisories[vulnID]; existsAlready {
-				return advisory.Advisories{}, fmt.Errorf("advisory already exists for %s", vulnID)
+				return v1.Advisories{}, fmt.Errorf("advisory already exists for %s", vulnID)
 			}
 
 			advisories[vulnID] = append(advisories[vulnID], advisoryEntry)
@@ -50,14 +50,14 @@ func Create(req Request, opts CreateOptions) error {
 	return fmt.Errorf("cannot create advisory: found %d advisory documents for package %q", count, req.Package)
 }
 
-func createAdvisoryConfig(cfgs *configs.Index[advisory.Document], req Request) error {
-	advisories := make(advisory.Advisories)
+func createAdvisoryConfig(cfgs *configs.Index[v1.Document], req Request) error {
+	advisories := make(v1.Advisories)
 
 	vulnID := req.Vulnerability
 	advisories[vulnID] = append(advisories[vulnID], req.toAdvisoryEntry())
 
-	err := cfgs.Create(fmt.Sprintf("%s.advisories.yaml", req.Package), advisory.Document{
-		Package: advisory.Package{
+	err := cfgs.Create(fmt.Sprintf("%s.advisories.yaml", req.Package), v1.Document{
+		Package: v1.Package{
 			Name: req.Package,
 		},
 		Advisories: advisories,
