+++ pkg/configs/advisory/v1/section_updaters.go
@@ -1,4 +1,4 @@
-package advisory
+package v1
 
 import (
 	"github.com/wolfi-dev/wolfictl/pkg/configs"
